Move bake command under Command/Bake.
